Wrap history embeddings search result timeouts with a Promise
This CL wraps the timeout that is forced to show the loading state for
a minimum time of 300ms with a Promise so that subsequent search results
for the same query can be chained to it. This ensures that the search
results are processed in correct order, instead of relying on the
correct timeout value.

Also updates hiding the answer section for when the search query is
changed and when there isn't a search result yet.

Bug: 371049023
Change-Id: I14b93cecdf28473accb1294d766b608aac8509f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5906521
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363909}