Ensure that ArCoreHostDispalyClient methods run on main thread
The `UseFrameIntervalDecider` feature seems to be triggering several of
the `viz::HostDisplayClient` messages on `ArCoreHostDisplayClient` for
the first time, which is exposing an error condition that these methods
are attempting to modify a View from a different thread than the one
responsible for it.

While it is unclear if we actually need these methods, this change
maintains the ability to call them by plumbing the main thread task
runner to the `ArCoreHostDisplayClient` and posting any needed calls
to the appropriate thread to avoid the crash.

Bug: 371244744
Change-Id: I0cb64ac255adf25bf7718ab7ef93f1a9ef65bc5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5908494
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363913}