Support android content-URIs via drop with ContentURLLoader
When a file is dropped on the toolbar, or I think also on any page
without a drop handler, the renderer loads the file as a navigation to a
file: URL.

For android content-URIs, we will use the existing content-URI rather
than wrap it inside a file: URL. This requires changes in a few places
where the file is handled and registered to allow navigation to it.

CompositorViewHolder.java: do not release the permissions to read a file
when a tab changes since a new tab is created for the navigation when a
file is dropped and we need to keep permission for the new tab.

browser_file_system_helper, child_process_security_policy_impl,
file_metadata: add support for files represented as content: URLs rather
than only file: URLs.

file_data_source: android content-URIs do not support seek() or pread(),
so use read() when possible, which is always the case for reading
a file after a drop.

Bug: 40101963
Bug: 363217587
Change-Id: Ia59ee6679512d1bf1a45eb412ecaeaa0675b6923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5832588
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363924}