Update "Open In Chrome" button logic to include non-eligible clients.
The recently-added "open in chrome" toolbar button (served on CCTs)
was built with the idea of being independent entity, but its
availability was linked to whether the embedder is allowed to request
interactive Omnibox in CCTs.

This change proposes a minor update, allowing clients that are not
eligible to request Omnibox in CCT to still request the "Open in
Chrome" toolbar button.

Bug: b/356157981
Change-Id: Ifcdd175ee9ba48b23905e75df506da87e52b441a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5906261
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Auto-Submit: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1363968}