Make the AI session creation wait for model download and installation
It's actually possible for the `GetOnDeviceModelStatus()` to return
kReady while the `model_metadata_` is not set yet. This happens when
the model is just being downloaded. This CL fixes the switch statement
and update the `kWaitableReasons` so the session creation can wait for
the model download and installation.


Bug: 361537114
Change-Id: I5d6fcc150eb01c5c4812dc786abbf0d24cf1fafd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5906964
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364022}