[Clipboard] Add CFSTR_INETURL format to clipboard on windows when image has non-empty source url.
The CL addresses the bug where browser does not add CFSTR_INETURL
or UniformResourceLocatorW format when image has empty source url
or empty alternate text (title). When right click is performed on
an image and copy image is triggered via context menu, the
|WriteImageToClipboard| in system clipboard is executed. This
calls |WriteImageWithTag| which in turn calls |WriteBookmark| and
redirects to the method in browser clipboard via mojom IPC. Here,
based on the condition that either source url or title is empty,
it drops further call and makes early return.

But, if the source url is non-empty, it should be added irrespective
of empty alternate text. The bug is fixed by removing the check on
alt text and only the url check is done for adding or not adding the
required format. This behavior is specific to Windows/Android, where
the format added is defined for Windows. In Android, the format is
already being added (Android specific) with or without the change.
MAC OS doesn't have this format and also for linux we never call
|ReadBookmark| and it is in NOT IMPLEMENTED state.

Bug: 40686999
Change-Id: Iaf9cbbf724fc9d8608537ee0c457b814f0225389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775430
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Pranav Modi <pranavmodi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1364042}