Reland: createImageBitmap: Remove more functions
Change the ImageBitmap constructor using ImageData to use
ApplyTransformsFromOptions.

As a result, FlipImageVertically, ScaleImage, and BakeOrientation are
no longer used. Remove them.

There is also only one caller to CreateProvider, so inline it, and
remove the now-dead branches.

Originally landed in crrev.com/1363357. This was reverted because
it preserved the RGBA ordering of ImageData, which SwiftShader
didn't support. Change the software path to always convert to N32,
like the blit path. The whole point of this work is to enable us to
have a single place to add F16 support.

Bug: 362431889
Change-Id: Ib55d431b8d2ea20229f9ab2389ac777339483470
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904027
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364046}