Fix another dangling pointer in PersistedTabDataAndroid::From()
PersistedTabAndroidData posts a UI task to itself to invoke
`FromCallback`. Unfortuantely, it's possible for the `TabAndroid` to be
destroyed by the time the posted task runs, which also destroys the
`PersistedTabDataAndroid`. Fix this by repeating the lookup in the
callback if the TabAndroid object is still alive.

An alternate fix is to add WeakPtrFactory to PersistedTabDataAndroid;
however, that is a bit suboptimal since it is preferable to embed
WeakPtrFactory in leaf classes. Otherwise, it is possible to end up in
a state where a WeakPtr is still valid even though Derived's destructor
has already run.

Bug: 365775467
Change-Id: I0099fa6d254ad91ff7842e6c3c027cb9456ed1ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5908424
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364048}