lobster: fix crash when re-creating the lobster session
When re-creating a new lobster session, both the old and new lobster
session will trigger `LobsterService::SetActiveSession` via
LobsterClient. Unfortunately, the old session is destroyed after the
creation of the new session. Therefore, the service holds a nullptr
value of the session instead of the new session. This causes crash when
the lobster session is created for the second time.

This CL does the following to fix this issue:
* Upon creating a new session, the session no longer asks the client to
`SetActiveSession` within the session's constructor. It also does not
ask the client to clear the active session when the session is destroyed

* We directly notify the lobster service about updating the active
session via lobster client at the time when the new session is created.

Bug: b:371484317
Change-Id: Icd695caee50d43c54828c7418358176cc7b9e0a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904584
Commit-Queue: Chuong Ho <hdchuong@chromium.org>
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364067}