arc: Add ArcUnthrottleOnActiveAudioV2 flag
The ArcUnthrottleOnActiveAudioV2 flag prevents ARC from being throttled
while there is active audio from ARC.

We launched the ArcUnthrottleOnActiveAudio feature, but we found a
performance regression that we want to fix. Since the original flag is
expired, we rename it to V2 (this is the same as removing the old flag
and creating a new V2 flag).

ArcUnthrottleOnActiveAudioV2's behavior is identical to the previous
version, but it is disabled by default. Future improvements and
experiments will be done with the V2 flag.

We decided not to implement a killswitch for older milestones because
the regression is not serious.

BUG=b:360924305, b:359337205
TEST=1. Set #arc-unthrottle-on-active-audio-v2 flag to default
     2. Play audio from Spotify in background
     3. Check /sys/fs/cgroup/cpu/arcvm/cpu.shares is 64
TEST=1. Enable #arc-unthrottle-on-active-audio-v2 flag
     2. Play audio from Spotify in background
     3. Check /sys/fs/cgroup/cpu/arcvm/cpu.shares is 1024

Change-Id: I454800e97157a810273118e2e7b911a354cde2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5898492
Reviewed-by: Yury Khmel <khmel@chromium.org>
Commit-Queue: Pattara Teerapong <pteerapong@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364073}