components: default the user-defined empty ctor and dtor where possible
I need an owner's approval please for changes under:
['components/content_settings', 'components/permissions']

This is a mechanical change. There are a few instances (but not all)
of empty user-defined constructors and destructors which can be
defaulted in the /components code.

This CL was uploaded by git cl split.

R=engedy@chromium.org

Bug: 371316188
Change-Id: I52a05e397b2419048ad4f811a03d628277bcf957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5906889
Auto-Submit: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364083}