[DBSC] Send an extra rotation request after session is terminated
This extra request communicates to the server the session termination
reason, needed for debugging.

The final request is made outside of the controller so that we don't
have to worry about zombie controllers. Controller is still removed as
soon as the session is terminated.

The extra rotation request is sent via a separate implementation of the
rotation fetcher class. This was done to avoid adding more complexity to
the existing fetcher as this would require the support of the send-only
mode with UMA off. This results in some code duplication but should be
less complex overall.

Fixed: 369566638
Change-Id: I87ac8b7e33eef19b10bdc688ca4d3cb9cee626c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904889
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364087}