Keep color picker views in sync with chosen color
This CL addresses two issues.
The first being that when the dialog is first shown, if the user opens
the advanced view, all sliders default to 0. This causes them to display
black despite the chosen color initially being red.

Secondly, although choosing a new color in the advanced view does update
the chosen color in the suggestions view, the reverse is not true. This
CL adds a call to setColor which means that when you choose a new color
in the suggestions view, the sliders will now update accordingly in the
advanced view.

Bug: 352690893
Change-Id: Ia6bc3729f80225d830795cd5fdd339f70ec85dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5905359
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Auto-Submit: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364114}