[DataSharing] Rename profile_dir -> data_sharing_dir in GroupDataModel
Pure rename with actual change in mind: it is probably not a good idea
to store DB directly in the profile directory, especially if we end up
having another DB (for favicons), but probably even without that.

Assuming this is not a GroupDataModel responsibility to create such dir
simply rename the parameter to make this expectation from the owning
class clear.

Bug: 370897286
Change-Id: I89b8bdbd71a5294956d8c203e8f6d1e51fc8d4d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5905499
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/main@{#1364121}