[Autofill] Update the current value of cached FormStructures
With the AutofillFixValueSemantics experiment, AutofillField exposes
both the field's current and the initial value.

It stands to reason that as a result, FormStructures obtained from
AutofillManager's form cache contain the fields' latest current
values. But that's not necessarily the case because AutofillManager's
events only update the cache if they reparse the form.

This CL fixes this: now, if AutofillFixValueSemantics is enabled,
cached FormStructures contain the latest FormData information the
browser process has received, including the latest current value.

All changes are behind the kAutofillFixValueSemantics feature.

Bug: 40227496, 40232021
Change-Id: Ie00cd1e466a6412077b8b38e9cafbea36a292b87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904422
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364152}