[sync] Simplify enum DataTypeConfigState
Refactoring without behavior changes, beyond the removal of DCHECKs.

The various "disabled" cases lead to identical behavior, so it is
pointless to distinguish different enum values and understand how to
precisely draw a line between them.

The non-trivial change in this patch is about the purging logic, where
unready types are no longer excluded. This should be fine, because
NIGORI never becomes unready (as it doesn't use a controller managed by
DataTypeManager). This also means the implementation in
BuildDataTypeConfigStateMap() can be simplified, because populating a
merged DISABLED bucket can simply be done via GetFailedTypes(), which is
already exercised before this patch (but is now made more explicit with
a CHECK).

Change-Id: Ifbd1f180e574e44bf9495df55e82d9220faf76b6
Bug: 40901755, 40154783
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5905321
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364154}