[sync] Untangle Nigori migration logic from main codepath
DataTypeConfigurer::ConfigureParams::to_purge, removed in this patch,
was a remnant of the old days when the sync engine had responsibilities
over cleaning up the Directory (original Sync architecture).

Over time, this logic was simplified, until it became exclusively needed
for (and supported for) Nigori. The reason why Nigori is different in
this context is that its controller is deep inside the sync engine and
not managed by DataTypeManager.

In addition, Nigori is a control type and therefore always enabled: the
only case where purging logic is needed is when dealing with the
protocol code MIGRATION_DONE, flagged by the Sync server to ask the
client to throw away all sync metadata.

In this patch, the migration logic for Nigori is refactored to decouple
it from the main codepath, removing oddities like
ConfigureParams::to_purge pretending to support arbitrary datatypes.

Change-Id: I434d03f75786089eb4f1c5112a78a7dd3d89c693
Bug: 40154783
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5909368
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364247}