[ios] List content of generated .storyboardc bundles
The output of compiling .storyboard files is a bundle with multiple
files so update the build rules to list all the generated files. It
should allow `gn clean_stale` to remove the individual files and it
would have prevented recent regression where a missing deps causing
the app to crash on startup was not failing for incremental builds.

    Note: the names of the generated files in the bundle seems to
    depend on the name of the .storyboard file. The algorithm is
    not known, but appears to be stable. For the moment the names
    have been hard-coded, it may be interesting to see if they can
    be automatically determined.

    When adding a new `bundle_data_storyboard_file(...)` target you
    can use `bundle_files = [ "Info.plist" ]` locally, run the build,
    look at the generated files, and update the target definition.

Bug: 371540753
Change-Id: Ib8e73c7a599b6384f2626220236660fc387179da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904165
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364249}