Avoid creating PdfOcrController for non-signed-in profile on ChromeOS.
ScreenAI service is created per profile and PdfOcrController (which uses
the service) is initialized during profile initialization, to be ready
to trigger the service for inaccessible PDFs.

On ChromeOS, there are two separate profiles for pre-sign-in and
post-sign-in and if screen reader is enabled, creating PdfOcrController
triggers the ScreenAI service initialization for both these profiles.

However the pre-sign-in profile cannot navigate to PDFs and hence does
not require this factory and service.

Bug: 371506467
Change-Id: I262e1e89909217d56c29ccdae31cb655c9ea95c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904123
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364265}