Make IsAcceptableSuggestionType() guard consistent in AutofillPopupControllerImpl.
This is a follow-up fix from discussion here:
https://crrev.com/c/5589627/comment/7f04ed3c_42649f98/
It does 2 things:
* makes desynchronization consistent (checking the size and
`IsAcceptableSuggestionType()`) in `AutofillPopupControllerImpl` and
`AutofillKeyboardAccessoryControllerImpl`
* adds a `CHECK` for desynchronization in `AcceptSuggestion()` as it is
done in `SelectSuggestion()`. This is legitimate because the only known
reason the check was "soft" in `AcceptSuggestion()` is the desync bug
(crbug.com/521133), which fix moved to the KA controller.


Fixed: 344814085
Change-Id: I33f7cc7ceba2222cd465c1da6bfbe27613b51c04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5804330
Commit-Queue: Dmitry Vykochko <vykochko@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1364292}