Apply mixins from a matrix config even if no variants are specified.
The targets.legacy_matrix_config declaration does not require that any
variants are specified, but in that case the matrix config ends up being
entirely ignored: the basic suite is expanded without any variants and
without incorporating any mixins that are specified in the matrix
config. This behavior is surprising, so this changes it so that the
mixins will be applied to the non-variant expansion of the test in the
basic suite.

Bug: 371449681
Change-Id: I3a2ba7783220ccb22d45665ab0066d58a0eb3a38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5909979
Reviewed-by: Marco Georgaklis <mgeorgaklis@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1364329}