[views-ax] Migrating a11y attributes in ui/views directory to new system
This ensures that the properties are updated in the accessibility cache whenever the underlying condition changes or the view is constructed. Additionally, the CheckedState, KeyShortcuts attributes are propagated to submenu_view and menu_scroll_view_container, along with the migration of kName and HasPopup.

The scroll_view_container class contains submenu_view, and submenu_view is associated with the root of menu_item_view. Actions performed on menu_item_view are propagated to submenu_view and scroll_view_container because references to these are present within menu_item_view. This ensures that any updates to attributes in menu_item_view are reflected in submenu_view and scroll_view_container.

In the MenuItemView::HandleAccessibleAction function, a new condition is being implemented to ensure that if menu_controller is null, a default action is executed. This is to address scenarios where actions might be performed without verifying the state of menu_controller.

The remaining instances will be migrated in follow-up CLs.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I944f5ae978c8590ab1d46130f21ec50b416acb60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5754645
Commit-Queue: Sejal Anand <sejalanand@microsoft.com>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1364342}