IDB: remove a lot of thread yielding when reading blob files
This change removes a substantial amount of indirection in both code
and task execution.

net::FileStream is replaced by base::File, which has the effect of
making i/o synchronous and same-thread. This should theoretically be
faster due to fewer deferrals and no longer relying on the possibly
heavily contended browser i/o thread. In this case, the calling
thread is the IDB bucket thread, which is the same thread that the
LDB database disk i/o is performed on.

Previously, `FileStreamReaderToDataPipe` was constructed on the
browser I/O thread (deferral #1), disk i/o occurred on a
threadpool taskrunner and reported back to the i/o thread (deferral
#2,3).

This is also a win for code simplicity since `base::File` is simpler
than passing through the layers of `storage::FileStreamReader`,
`net::FileStream` and `net::FileStream::Context`.

I did consider also running each blob reader on a different task
runner and not using the bucket thread, however that didn't benchmark
any differently, so I decided against the added complexity.

*** Benchmarking ***

I updated the IDB playground to better test this:
https://github.com/evanstade/indexeddb_playground/commit/c13325d961739978e9850325554fe86e5d79509f

With 100B blobs, there is no consistently measurable performance
impact on two of the three tests. On the third, "getting each item
in its own transaction", there was an improvement of approximately
10% which was pretty consistent.

When increasing the blob size to 1MB, one benchmark was unimproved,
one was improved by about 7%, and the "blob per transaction" test
was improved by 25%.

These results are pretty load-specific and device dependent, but even
in cases where this change has little measured performance impact, it
has a theoretical benefit which is harder to measure: to reduce
dependence on (and contention of) the I/O thread, which would be busy
for reasons unrelated to IDB on systems experiencing high load.

Bug: 40262766,40518562,369319575
Change-Id: I9ec1dc5d5462d0ebdbd08d3a7c5b0b92642db6df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5898304
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Brad Triebwasser <btriebw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364376}