[lensoverlay] Add gsessionid, vsrid, and vsint to suggest requests
These are passed via the LensOverlaySuggestInputs (which I renamed
from LensOverlayInteractionResponse). The logic changes to the overlay
controller will be in a followup cl. In the future, some suggest inputs
will be available before a Lens interaction, so the name needed to be
updated.

Bug: 367404756, 370987524
Change-Id: Id1e6373a2ac4f8e65b7a1117d6daca4ca0b9976d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5902495
Reviewed-by: Christian Xu <christianxu@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1364397}