Protected Audiences: Introduce TrustedSignalsKVv2Manager.
This class manages requests for trusted signals KVv2 from the
Javascript process to the TrustedSignalsCache in the main process.
Unlike the legacy TrustedSignalsRequestManager, it's per-process rather
than per-worklet, so should work better when multiple scripts from the
same origin use the same signals URL.

It merges requests for the same compression group into a single request
to the TrustedSignalsCache to minimize redundant transferring and
parsing of data.

It also caches responses - when a compression group is requested, it
transfers the group from the browser process, parses the group in its
entirety, and keeps the parsed result in memory until there are no live
requests for any partition within the compression group. The intention
is for the consumer to hold onto the request until its run the V8
script that needs is, so when there are a bunch of pending V8 calls
to a script all sent out by the browser at once, the data is likely
to be reused for all requests, regardless of when the data is received
from the cache.

BYPASS_LARGE_CHANGE_WARNING

Bug: 333445540
Change-Id: Ifa95a8c110a1940ef9d956dc5c21df266ece9d5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5902518
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364445}